cleanups

+++ networkb/wfile_transfer_file.cpp
@@ -42,7 +42,7 @@ bool WFileTransferFile::GetChunk(char* chunk, size_t start, size_t size) {
     }
   }
   
-  if ((start + size) > file_size()) {
+  if (static_cast<int>(start + size) > file_size()) {
     clog << "ERROR WFileTransferFile::GetChunk (start + size) > file_size():"
          << "values[ start: " << start << "; size: " << size
 	 << "; file_size(): " << file_size() << " ]" << endl;
